refactoring: set constraints & business logic for employee, department

+++ staffflow-server/src/main/java/kz/zzhalelov/staffflow/server/service/impl/DepartmentServiceImpl.java
@@ -1,7 +1,10 @@
 package kz.zzhalelov.staffflow.server.service.impl;
 
+import kz.zzhalelov.staffflow.server.exception.NotFoundException;
 import kz.zzhalelov.staffflow.server.model.Department;
+import kz.zzhalelov.staffflow.server.model.Employee;
 import kz.zzhalelov.staffflow.server.repository.DepartmentRepository;
+import kz.zzhalelov.staffflow.server.repository.EmployeeRepository;
 import kz.zzhalelov.staffflow.server.service.DepartmentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,9 +15,14 @@ import java.util.List;
 @RequiredArgsConstructor
 public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentRepository departmentRepository;
+    private final EmployeeRepository employeeRepository;
 
     @Override
     public Department create(Department department) {
+        Long managerId = department.getManager().getId();
+        Employee manager = employeeRepository.findById(managerId)
+                .orElseThrow(() -> new NotFoundException("Сотрудник не найден"));
+        department.setManager(manager);
         return departmentRepository.save(department);
     }
 
@@ -25,7 +33,8 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public Department findById(long departmentId) {
-        return departmentRepository.findById(departmentId).orElseThrow();
+        return departmentRepository.findById(departmentId)
+                .orElseThrow(() -> new NotFoundException("Отдел не найден"));
     }
 
     @Override
@@ -35,6 +44,10 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public void delete(long departmentId) {
-        departmentRepository.deleteById(departmentId);
+        if (departmentRepository.findById(departmentId).isPresent()) {
+            departmentRepository.deleteById(departmentId);
+        } else {
+            throw new NotFoundException("Отдел не найден");
+        }
     }
 }
